--- src/components/common.tsx
@@ -1,4 +1,0 @@
-export type EthereumProvider = {
-    request(...args: unknown[]): Promise<unknown>,
-    on(e: string, handler: (x: string) => void): void,
-}
